Resolve most TypeScript errors
