+++ packages/php-wasm/src/web/spawn-worker-thread.ts
