+++ packages/php-wasm/tsconfig.json
@@ -1,18 +1,44 @@
 {
-	"extends": "./tsconfig.base.json",
 	"compilerOptions": {
-		"lib": [ "dom", "ESNext" ],
+		"allowJs": true,
+		"checkJs": false,
+		"allowSyntheticDefaultImports": true,
+
+		"jsx": "react",
+
+		"target": "esnext",
+		"module": "esnext",
+		"declaration": true,
+		"declarationMap": false,
+		"composite": true,
+		"emitDeclarationOnly": true,
+		"isolatedModules": true,
+		"removeComments": false,
+        "skipLibCheck": true,
+
+		"strict": true,
+
+		"noImplicitAny": false,
+		"noUnusedLocals": true,
+		"noUnusedParameters": false,
+		"noImplicitReturns": true,
+		"noFallthroughCasesInSwitch": true,
+
+		"moduleResolution": "bundler",
+
+		"esModuleInterop": false,
+
+		"rootDir": "src",
 		"declarationDir": "build/web/types",
-		"outDir": "build/web/types"
+		"outDir": "build/web/types",
+		"baseUrl": ".",
+		"lib": [ "dom", "ESNext" ]
 	},
-	"include": [ "src/**/*" ],
+	"include": [ "src/" ],
 	"exclude": [
-		 "src/web/service-worker/worker-library.ts",
-		 "src/**/*.node.js", 
-		 "src/**/*.node.ts", 
-		 "src/php-compilation/**/*", 
+		 "src/php-build", 
 		 "src/__tests__/*.ts",
-		 "build/**/*",
-		 "build-*/**/*",
+		 "build",
+		 "build-*",
  	]
 }
