--- packages/php-wasm/tsconfig.node.json
@@ -1,8 +1,0 @@
-{
-	"extends": "./tsconfig.base.json",
-	"compilerOptions": {
-		"declarationDir": "build/node/types",
-		"outDir": "build/node/types",
-	},
-	"include": [ "src/php-library/*", "src/php/*.node.js" ]
-}
