--- packages/php-wasm/tsconfig.worker.json
@@ -1,24 +1,0 @@
-{
-	"extends": "./tsconfig.base.json",
-	"compilerOptions": {
-		"lib": ["WebWorker"],
-		"declarationDir": "build/web/types",
-		"outDir": "build/web/types"
-	},
-	"include": [
-		"src/urls.ts",
-
-		"src/php-library/index.ts",
-		"src/php-library/php.ts",
-		"src/php-library/utils.ts",
-		"src/php-library/php-server.ts",
-		"src/php-library/php-browser.ts",
-		 
-		"src/web/service-worker/index.ts",
-		"src/web/service-worker/worker-library.ts",
-		"src/web/emscripten-download-monitor.ts",
-		"src/web/messaging.ts",
-		"src/web/scope.ts",
-		"src/web/utils.ts"
- 	]
-}
