--- packages/wordpress-playground/tsconfig.base.json
@@ -1,37 +1,0 @@
-{
-	"compilerOptions": {
-		"allowJs": true,
-		"checkJs": false,
-		"allowSyntheticDefaultImports": true,
-
-		"jsx": "react",
-
-		"target": "esnext",
-		"module": "esnext",
-		"declaration": true,
-		"declarationMap": false,
-		"emitDeclarationOnly": true,
-		"isolatedModules": true,
-		"removeComments": false,
-		"skipLibCheck": true,
-
-		"strict": true,
-		"composite": true,
-
-		"noImplicitAny": false,
-		"noUnusedLocals": true,
-		"noUnusedParameters": false,
-		"noImplicitReturns": true,
-		"noFallthroughCasesInSwitch": true,
-
-		"moduleResolution": "bundler",
-
-		"esModuleInterop": false,
-
-		"declarationDir": "build/types",
-		"rootDir": "src",
-		"outDir": "build/types",
-		"baseUrl": "."
-	},
-	"references": [{ "path": "../wordpress-plugin-ide/tsconfig.json" }]
-}
