--- packages/wordpress-playground/tsconfig.worker.json
@@ -1,15 +1,0 @@
-{
-	"extends": "./tsconfig.base.json",
-	"compilerOptions": {
-		"lib": ["WebWorker"]
-	},
-	"include": [
-		 "src/service-worker.ts",
-		 "src/worker-thread.ts",
-		 "src/worker-utils.ts",
-		 "src/config.ts"
- 	],
-	 "references": [
-		 { "path": "../php-wasm/tsconfig.worker.json" }
-	 ]
-}
